requirements.txt
asgiref==3.2.7
beautifulsoup4==4.8.2
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.5
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-form==3.4
django-bootstrap4==1.1.1
python-decouple==3.3
pytz==2019.3
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
static3==0.7.0
gunicorn
psycopg2
